add vector, simplify deployment files
